Reorganize data component files (w/ Claude)
