+++ src/data/CMakeLists.txt
@@ -1,11 +1,20 @@
 #
 #  This file is part of Fenix
-#  Copyright (c) 2016 Rutgers University and  Sandia Corporation.
+#  Copyright (c) 2016 Rutgers University and Sandia Corporation.
 #  This software is distributed under the BSD License.
 #  Under the terms of Contract DE-AC04-94AL85000 with Sandia Corporation,
 #  the U.S. Government retains certain rights in this software.
 #  For more information, see the LICENSE file in the top Fenix
 #  directory.
 #
+
+target_sources(fenix PRIVATE
+    recovery.cpp
+    component.cpp
+    group.cpp
+    policy_imr.cpp
+    member.cpp
+    snapshot.cpp
+    subset.cpp)
 
 add_subdirectory(util)
