+++ test/subset/subset_pack_data.cpp
@@ -56,6 +56,7 @@
 */
 
 #include <fenix.h>
+using fenix::data::util::DataBuffer;
 #include <mpi.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -63,9 +64,12 @@
 #include <sys/types.h>
 #include <unistd.h>
 
-#include <fenix_data_subset.hpp>
-#include <fenix_data_member.hpp>
+#include "fenix/data/subset.hpp"
+using fenix::data::util::DataBuffer;
+#include "fenix/data/member.hpp"
+using fenix::data::util::DataBuffer;
 #include <fenix/data/util/serializer.hpp>
+using fenix::data::util::DataBuffer;
 
 #include "subset_common.hpp"
